Monotonicity state moved into `GenerationOptions` instance. Identical performance.
